requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
requests
psycopg2-binary
redis
neo4j
pgvector
numpy
openai
tenacity
sentry-sdk
python-telegram-bot
